move fixtures folder
