+++ src/api/lib/expect/index.ts
@@ -1,0 +1,2 @@
+export { expect as statusExpect } from './statusExpect';
+export { expect as typeExpect } from './typeExpect';
