+++ src/fixtures/baseAPIFixture.ts
@@ -1,6 +1,6 @@
 import { mergeExpects } from '@playwright/test';
-import { expect as typesExpect } from 'src/api/fixtures/typesExpect';
-import { expect as statusesExpect } from 'src/api/fixtures/statusesExpect';
+import { expect as typesExpect } from '@apiExpect/typeExpect';
+import { expect as statusesExpect } from '@apiExpect/statusExpect';
 
 export { test } from '@playwright/test';
