--- src/pages/fixtures/stringsExpect.ts
@@ -1,7 +1,0 @@
-// import { test as baseTest } from "@playwright/test";
-
-// export const test = baseTest.extend({
-//   toHaveTrimmedText(received: any, value) {
-//     const isMatchesText = received.trim
-//   }
-// });
